requirements.txt
requests==2.25.1
urllib3==1.26.4
chardet==4.0.0
certifi==2020.12.5
idna==2.10
python-telegram-bot
lxml
arrow
